Added headlight support to client
Added check for 'select' button to step headlight value from 0 to max and around again.  Default values take it from 0->34->68->100->0.  Other values are supported by changing the constants prefixed with HEADLIGHT.

This requires support in the Arduino code to listen for "Lnnn" events and set the PWM of a pin from the value (or to implement a threshold to activate/deactivate a digital pin).